Add option to disable sending module info
Users can now opt-out of automatically sending this data by setting the
config attribute `send_modules` to `false`.

In projects which version their dependencies using a Gemfile, sending
the list of loaded modules may not provide much value. In particular, if
you send a parameter including what version of the application is
running, we can inspect the Gemfile.lock for all this data, and don't
need it to appear in every sentry instance.