Refactor how Rack exceptions are captured
The interface is Raven::Rack.capture_exception. The implementation
uses the new yield behavior of Raven.capture_exception to avoid
needing a special-case capture_rack_exception on Event or to
manually send the event.

For parallelism, define Raven::Rack.capture_message as well. See
the next commit for a use case.