Change `Rails.inject` to look for the `railties` gem, not `rails`.
`rails` is a meta gem that can be skipped in a Rails app if we don't want to
install gems like `actionmailer`, `activerecord` or `activejob`, while the
`railties` gem is always present in a Rails app.

To ensure that Raven is properly loaded when the `rails` gem is not present
we look for the `railties` gem, not `rails` in the `Gem.loaded_specs` Hash.