Check catch_debugged_exceptions option at runtime
Instead of adding the hook to report debugged exceptions based on the
value of the configuration option while Rails is initialising, we can
always add the hook, but have it check the option at runtime. This is
more in line with how other configuration options are used, and means
that the option can be set or changed at any time.

This also lets us write tests that exercise this behaviour, since the
logic is no longer tied to the Rails initialisation process which can
only happen once during the test suite.