Return original response in debug middleware patch
The call to the original render_exception method was moved into an
ensure block to prevent errors in this module from interfering with
Rails' own error handling behaviour. However an ensure block has no
effect on the return value of the method, and we were returning the
result of Raven::Rack.capture_exception instead of the response
generated by the original Rails middleware.