Always match excluded exceptions by case equality
... no matter if they are defined by string or constant in
Raven.configuration[:excluded_exceptions].

The upside of doing it like this is that Raven will have the same match
behavior no matter how the exception was excluded. Previously a
string-defined excluded exception wouldn't match inherited/extended
exceptions.

Use case example:

  # Raven setup
  Raven.configure do |config|
    config.excluded_exceptions = ["SentryExcludedException"]
  end

  # Application
  module SentryExcludedException; end

  # The application can decide what to exclude
  begin
    raise StandardError, "oh snap"
  rescue => error
    error.extend SentryExcludedException
    raise
  end

This only worked if the excluded_exceptions was configured with
constants, but not with strings, which this fixes.

By using strings the coupling between the Raven setup/environment
and the application is reduced, improving unit testing etc.