Fix wrong Raven.user_context usage breaking event sending
When using Raven.user_context with `options` argument that's not a Hash,
the Raven::Event#initialize method will raise an Exception
and prevent an event being sent to Sentry at all.

Of course, this is not acceptable for an exception-sending service,
therefore log a warning that the Raven.user_context `options` argument
will be ignored to ensure that an event will be sent to Sentry.