Allow `current_environment` to be set via an env variable
This commit allows developers to set a `SENTRY_CURRENT_ENV` variable to
configure the `current_environment` setting.

It serves as a fix for this scenario:

We run our production/staging/integration environments with Rails env
"production". To differentiate errors from the environments we set
`config.current_environment` with an environment variable
(https://github.com/alphagov/content-tagger/blob/bac2e1903aaf6ac12b3a103
490517355673648dd/config/initializers/sentry.rb#L2).

However, this doesn't work in all situations. For example, when you run
a rake task that doesn't exist, an error will be reported to Sentry.
Because the initialiser that sets the `current_environment` hasn't run
(yet) Sentry will use the `RAILS_ENV` to determine the current env.
This causes all errors to be reported as production instead of
production/staging/integration.