Add `after_send_event` post send hook
This adds a hook that is called after an event is successfully sent to
Sentry.

It is mutually exclusive with `transport_failure_callback`. Together
with that hook this fulfills the "Post-send hook" feature in the list
of [expected features in clients][1]:

> Hook called on success or error, and passed the event and exception
so that users can perform actions based on events actually being sent.

I've looked in the other client libraries for a good name for this.
I've found 2 similar configs: `after_send_event` is used by the [elixir
library][2], and `send_callback` is used by [PHP][3]. I liked
`after_send_event` slightly better.

[1]: https://docs.sentry.io/clientdev/features/#post-send-hook
[2]: https://docs.sentry.io/clients/elixir/config/
[3]: https://docs.sentry.io/clients/php/config/

Closes https://github.com/getsentry/raven-ruby/issues/737.