Fix bugs with send_event in asynchronous mode (#940)
* Log event_id even when in asynchronous mode

If we're using Raven asynchronously, we call `to_json_compatible` on
event which results in all of the keys of the `event` hash being
Strings, not Symbols. So `event[:event_id]` will return `nil` and the
log won't have the `event_id`.

* Fix bug with get_log_message in asynchronous mode

When running asynchronously, `event` will be a hash with String keys,
unlike in synchronous mode where they're Symbols.