Make get_message_from_exception a public method
As of #928, `Event`s no longer expose a `message` attribute. For code
that relied on using `message`, it's cumbersome to now have to do this
whole nil-checking chain every time. It would be nice if instead we
could use the already existing `get_message_from_exception` method to
handle this for us.