Circular references processor should not change user data
I understand that it's not memory-efficient, but you are working with business rules which might rely on references.
Consider opting out this behavior, or display a warning somewhere, and this behavior should be documented. 
I spent 6 hours of debugging code because of the sentry event. Imagine my feelings when I figure out it's caused by raven.