Update README's Async Section
It's really not good to show a code example using Ruby threads and then immediately recommend to the majority of users (MRI/CRuby) that they should avoid using it. Many copy and paste and go about their day. I just removed that exact code snippet from a large app and switched it run in AR 🙃 Many people implementing an exception tracker, at least for the first time, could be very new to Ruby or programming, and not understand how much thread contention can degrade performance.

The better example was already there. That's all that is needed. I figured I'd toss this your way and see if you agree. Maybe in the future async could be handled internally by raven itself so the user doesn't have to concern themselves with the joys of mri threads 😅

Thanks for all the work put into this gem btw. I've never had any issues with raven over the many years of using it!