Support new sidekiq_report_type config option
Sidekiq takes 2 types of handlers to deal with failed jobs:
- error handler - which will be called every time a job failed
- death handler - which will only be called when the job is dead (no
  retries remained)

And currently, our sidekiq integration is based on the error handler. So
if a job failed repeatedly, the exception will be sent multiple times.

There have been some requests on "only sending events when there's no
retry left" and I think this can be a solution for them.