Remove Sentry::Rack.capture_exception
The method was used to inject request env into the Event object, but now we
already have the following workflow for that:

```
add a temp scope -> inject env to the temp scope -> apply to the event with other scope
attributes
```

we don't need this method anymore. Sentry.capture_exception should just
do the work.