ActiveSupportLogger should only record events that has a started time
Because `ActiveSupport::Notifications.publish` can be used to publish
anything, whether it's standard Rails instrumentation payload, a string,
or even a custom object, it's dangerous for the logger to accept anything
that's being published.

So this commit filters out events that are not compatible with the
default Rails instrumentation format to avoid issues.