fix: retry jobs that raise exceptions
CleanupMiddleware was swallowing exceptions raised in jobs, preventing
Sidekiq from adding the job to the retry queue. Re-raising from the
CleanupMiddleware led to a second firing from the ErrorHandler
class. To get around this, this commit narrows the scope of
ErrorHandler to only handle Sidekiq lifecycle events, and renames the
ErrorHandler class to EventErrorHandler accordingly.

Also fixes (I think) a second issue, where active job context key
renaming was not being applied on the CleanupMiddleware class, which
was the class responsible for firing job error events. Ensures that
the context filter is called in the CleanupMiddleware as well.