Copy request env before Rails' ShowExceptions middleware
When Rails' `ShowExceptions` middleware calls `#render_exception`, it
also updates some of the env's headers, like changing `PATH_INFO` to
`/500`.

And if the current scope holds the same env hash as ShowExceptions does, this
behavior will affect the reported event as well. Because
the CaptureExceptions middleware will always run after ShowExceptions
is called.

So the solution to copy current scope's rack env when
we know the exceptions app is going to be rendered.

Notes:
- sentry-raven patches ShowExceptions and reports events before
  `ShowExceptions#render_exception` is called, so it doesn't have such issue.
- `ShowExceptions#render_exception` only changes first-level env
  attributes. So a simple `dup` should be able to solve the issue, don't
  need to deep-copy the entire env hash.