Prevent exceptions app from overriding event's transaction name
Because exceptions app also works like a normal controller action, when
it renders the exception, it'll push a new transaction name to the scope
as well. But that's not the transaction name we want to display in
Sentry.

But unlike how we deal with the rack env overriding issue, this time we
need to use event processors to replace the transaction name.