Separate individual breadcrumb's data serialization
Breadcrumb can literally contain any type of data, and some of them can
cause problems when being serialized, like a hash or array that contains
circular references. Currently, this is causing the entire event to be
ditched because of the serialization error or SystemStackError, which is
bothering the users.

However, to stop bad data from being recorded is almost impossible, or at
least it'll be extremely expensive to check. So this commit takes a
different approach:

1. we use `JSON.parse(JSON.generate(breadcrumb_data))` in
   `Breadcrumb#to_hash` to make sure the generated Hash is
   serializable.
2. if a breadcrumb's data is not serializable, the data will be
   thrown away but the breadcrumb will be preserved.
3. whenever the serialization error happens, Sentry will log related
   messages to help future debugging.