ci(release): Fix changelog entry paths in craft config
Since we do per-directory releases in this repo, we need to tell Craft where the changelog for each project is (otherwise it just looks at the repo root). This is why our GitHub releases were lacking release notes. This PR also moves the GitHub target to the last to ensure all main targets are published before creating a GitHub release tag.