Fix YARD::Handlers::Ruby::MixinHandler in rake extension
```
[warn]: Load Order / Name Resolution Problem on Rake::Application:
-
Something is trying to call mixins on object Rake::Application before it has been recognized.
This error usually means that you need to modify the order in which you parse files
so that Rake::Application is parsed before methods or other objects attempt to access it.
-
YARD will recover from this error and continue to parse but you *may* have problems
with your generated documentation. You should probably fix this.
-

[error]: Unhandled exception in YARD::Handlers::Ruby::MixinHandler:
  in `lib/sentry/rake.rb`:31:

        31: Rake::Application.prepend(Sentry::Rake::Application)
```