Capture exceptions in Action Cable connections and channels (continuation of #1295) (#1638)
* capture exceptions in Action Cable connections and channels

There are 5 types of hooks that Action Cable provides:

* `Connection#connect`
* `Connection#disconnect`
* `Channel#subscribed`
* `Channel#unsubscribed`
* `Channel` actions

Now, any exceptions raised within those hooks are captured by Sentry, and reported as `ActionCable/[...]` transactions. Additional context is included depending on the hook the exception was raised within.

A note/quirk: the Rack env that's included in the scope is from the `Connection`, and therefore has a URL of the cable `mount_path` (usually `/cable`) as well as the headers from that initial connection request.

Additionally, there is not currently a really clean way to hook in and set `user_context`. I don't know if that is a blocker for this integration, but wanted to make sure it was noted.

* Only extend ActionCable when it's defined

Similar to the update for ActionCableExtensions in #1218 and #1494

* test(ActionCable): Assert only 1 event captured:

https://github.com/getsentry/sentry-ruby/pull/1295/#discussion_r584484063
https://github.com/getsentry/sentry-ruby/pull/1295/#discussion_r584484088

* Remove namespace 'ActionCable/*'

https://github.com/getsentry/sentry-ruby/pull/1295/#discussion_r584489606

* Test ActionCable #unsubscribed

https://github.com/getsentry/sentry-ruby/pull/1295/#discussion_r584484867

* test(ActionCable): prefer rspec-rails over mini-test

https://github.com/getsentry/sentry-ruby/pull/1295#discussion_r584483392

* test(ActionCable): cleanup duplicated set_callback

* docs(CHANGELOG): update #1295 to #1638

Co-authored-by: Alex Robbin <agrobbin@gmail.com>