Fix: Don't expect `.backtrace` to be present
If a non-`Error` type is passed to the `Raven.capture_exception` method,
for instance when ActiveRecord errors are returned, and we've set up
`backtrace`s to be recorded, Raven itself raises an exception because
the `.backtrace` isn't found.

By checking that `:backtrace` is present first, we can make sure we
don't hit these error conditions.