Replace sentry-ruby-core with sentry-ruby as integration dependency
sentry-ruby-core was created for loosing version requirement on faraday.
It's the actual code container and sentry-ruby just wraps around it with
a version requirement. But since faraday is now dropped, this extra
structure is not needed anymore.

The plan is to make sentry-ruby the gem that holds the logic and point
integrations to it. And sentry-ruby-core will just depend on sentry-ruby
and does nothing for backward compatibility.