bugfix: sentry crash when exeception message is an array.
## Problem

when the exception message is not a string, sentry will crash and
sentry sdk will cause main application crash either.

Here is an example to crash sentry.

For example

```
raise StandardError.new(message: ["msg1", "msg2"])
```

I know the developer should use a standard string as the exception
message, but not all developers are well educated. Sentry as a service
provider can't always assume the exception message is a string.

Sometimes it's a hash or array.

## Solution

if the exception message is not an string, we should not call byteslice
method.