Support capturing Puma's error
This PR implements Puma's error reporting by patching `Puma::Server#lowlevel_error`.

I thought about using the `lowlevel_error_handler` API to add a Sentry error
handler, but it's not possible to do that after `puma.rb` is loaded (which
is extremely early in a normal app setup, way before the SDK is loaded).
There is also no good ways to fetch and update the server's `lowlevel_error_handler` option.

So the only option left is to patch the method.