Improve default slug generation for `sidekiq-scheduler`
`sidekiq-scheduler` allows to use a class name as a schedule name directly as follows.

```
Namspeced::CancelAbandonedOrders:
  cron: '0 */5 * * * *'
```

In this case, a slug is shown as `namspecedcancelabandonedorders` on Sentry.
I think this isn't good for readability.
This PR converts `::` to `-` as well as the cron default behavior of Crons mixin.

Related with: #2138.