Add `bigdecimal` to the gemspec
Since Ruby 3.3.0, RubyGems and Bundler warn if users do require the
gems that will become the bundled gems in the future version of Ruby.

Please see the "Standard library updates" section for the details.
https://www.ruby-lang.org/en/news/2023/12/25/ruby-3-3-0-released/

`sentry-ruby` uses one of those gems, `bigdecimal`.
So now get the warning like the following.
https://github.com/getsentry/sentry-ruby/actions/runs/7463068254/job/20306920751#step:6:89

This PR adds `bigdecimal` to the gemspec to fix the warning.