Stop relying on fugit
It is an overkill to have a (implicit) dependency
on fugit, especially that it doesn't even offer
a public interface to do what we need.

It's older version don't support timezones in cron
which I believe is the culprit in #2513.

Even if it's not, I still think it's a good change.