Replace faulty spec with something more useful
Not sure what the intention was but this spec
did not test any faraday error handling. What
actually happened was that HTTPTransport really
tried to hit the API and ended up with error.

We *do not* want to hit any APIs like that in specs,
this causes instability of the spec suite.

That's why I added webmock to ensure we're not sending
any HTTP requests in specs.