Make DummyTransport#send_event call super
This is important because that's how actual Transports
behave.

The specs had to be changed because their expectations
were based on behavior specific to DummyTransport.

Reasons for changing spec expectations:

Log spec creates a transaction and we run profiler
for every transaction, but due to sampling decision
profile event is being dropped and we end up with
a client report.

SessionFlush calls capture_envelope which bypasses
send_event, but the test actually does send two
error events, so now that DummyTransport#send_event
calls super, we end up (correctly) with 3 envelopes:
2 from errors, and 1 from session.