feat: include otel as custom sampling contex (#2683)
* feat: include otel as custom sampling contex

Sentry's opentelemetry adapter does not do the greatest job in
adapting the opentelemetry span to a semantically correct sentry span,
nor does it expose the opentelemetry span in the context for sampling
decisions. This PR adds a opentelemetry span as an additional custom sampling
context so it can be used for sampling decisions. This gives clients more
opportunity to use the entire context of the opentelemetry span for sampling
decisions, and allows replicating sampling mechanisms as orignally suggested
by the Sentry team via the "Sampling Function".

https://docs.sentry.io/platforms/ruby/configuration/sampling/#setting-a-sampling-function

* chore: exercise otel context with tracer_sampler